refactor Plot.Layout()
